add worker pool with func

+++ utils/init_settings.go
@@ -25,7 +25,7 @@ type AppConfig struct {
 	IpVersion         string `mapstructure:"ip_version"`
 	MaxConn           int    `mapstructure:"max_conn"`
 	MaxPacketSize     uint32 `mapstructure:"max_packet_size"`
-	WorkerPoolSize    uint64 `mapstructure:"worker_pool_size"`
+	WorkerPoolSize    int    `mapstructure:"worker_pool_size"`
 	MaxWorkerTaskLen  uint64 `mapstructure:"max_worker_task_len"`
 	HeartbeatMaxTime  int    `mapstructure:"heartbeat_max_time"`
 	HeartbeatInterval int    `mapstructure:"heartbeat_interval"`
